move build-list.py
